// frontend/src/api/generated/models/user.ts
/* tslint:disable */
 
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * ユーザースキーマ（レスポンス用）
 * @export
 * @interface User
 */
export interface User {
    /**
     * ユーザー名
     * @type {string}
     * @memberof User
     */
    'username': string;
    /**
     * メールアドレス
     * @type {string}
     * @memberof User
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'id': string;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'is_active'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'is_verified'?: boolean;
    /**
     * ユーザーのロール一覧
     * @type {Array<string>}
     * @memberof User
     */
    'roles'?: Array<string>;
    /**
     * 
     * @type {Date}
     * @memberof User
     */
    'created_at': Date;
    /**
     * 
     * @type {Date}
     * @memberof User
     */
    'updated_at': Date;
}