// frontend/src/api/generated/models/character-stats.ts
/* tslint:disable */
 
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * キャラクターステータススキーマ
 * @export
 * @interface CharacterStats
 */
export interface CharacterStats {
    /**
     * レベル
     * @type {number}
     * @memberof CharacterStats
     */
    'level'?: number;
    /**
     * 経験値
     * @type {number}
     * @memberof CharacterStats
     */
    'experience'?: number;
    /**
     * 現在HP
     * @type {number}
     * @memberof CharacterStats
     */
    'health'?: number;
    /**
     * 最大HP
     * @type {number}
     * @memberof CharacterStats
     */
    'max_health'?: number;
    /**
     * 現在MP
     * @type {number}
     * @memberof CharacterStats
     */
    'mp'?: number;
    /**
     * 最大MP
     * @type {number}
     * @memberof CharacterStats
     */
    'max_mp'?: number;
    /**
     * 
     * @type {string}
     * @memberof CharacterStats
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof CharacterStats
     */
    'character_id': string;
}