// frontend/src/api/generated/models/narrative-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { ActionChoice } from './action-choice'
// May contain unused imports in some cases
// @ts-ignore
import type { LocationEvent } from './location-event'

/**
 * 物語生成のレスポンス
 * @export
 * @interface NarrativeResponse
 */
export interface NarrativeResponse {
  /**
   * 生成された物語テキスト
   * @type {string}
   * @memberof NarrativeResponse
   */
  narrative: string
  /**
   * 場所が変更されたか
   * @type {boolean}
   * @memberof NarrativeResponse
   */
  location_changed?: boolean
  /**
   *
   * @type {string}
   * @memberof NarrativeResponse
   */
  new_location_id?: string | null
  /**
   *
   * @type {string}
   * @memberof NarrativeResponse
   */
  new_location_name?: string | null
  /**
   * 消費されたSP
   * @type {number}
   * @memberof NarrativeResponse
   */
  sp_consumed?: number
  /**
   * 次の行動選択肢
   * @type {Array<ActionChoice>}
   * @memberof NarrativeResponse
   */
  action_choices: Array<ActionChoice>
  /**
   *
   * @type {Array<LocationEvent>}
   * @memberof NarrativeResponse
   */
  events?: Array<LocationEvent> | null
}